Split installation utilities into multiple files
