+++ pkg/installation/installation.go
@@ -1,11 +1,6 @@
 package installation
 
 import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/url"
-
 	"github.com/giantswarm/microerror"
 )
 
@@ -46,52 +41,4 @@ func New(fromUrl string) (*Installation, error) {
 	}
 
 	return i, nil
-}
-
-type installationResponse struct {
-	Installation installationInfo `json:"installation"`
-}
-
-type installationInfo struct {
-	Provider  string `json:"provider"`
-	K8sCaCert string `json:"k8s_cacert"`
-	Name      string `json:"name"`
-}
-
-func getInstallationInfo(apiUrl string) (installationInfo, error) {
-	res, err := http.Get(apiUrl) // #nosec G107
-	if err != nil {
-		return installationInfo{}, microerror.Mask(err)
-	}
-
-	defer res.Body.Close()
-
-	result := installationResponse{}
-	{
-		err = json.NewDecoder(res.Body).Decode(&result)
-		if err != nil {
-			return installationInfo{}, microerror.Maskf(cannotGetInstallationInfo, "make sure you're connected to the internet and behind a VPN")
-		}
-
-		result.Installation.K8sCaCert = parseCertificate(result.Installation.K8sCaCert)
-	}
-
-	return result.Installation, nil
-}
-
-func getBasePath(k8sAPIUrl string) (string, error) {
-	k8sURL, err := url.Parse(k8sAPIUrl)
-	if err != nil {
-		return "", microerror.Mask(err)
-	}
-
-	return k8sURL.Host, nil
-}
-
-func getGiantSwarmAPIUrl(basePath string) string {
-	return fmt.Sprintf("https://%s.%s", apiURLPrefix, basePath)
-}
-
-func getAuthUrl(basePath string) string {
-	return fmt.Sprintf("https://%s.%s", authURLPrefix, basePath)
 }
